requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comtypes==1.4.9
docopt==0.6.2
docstring_parser==0.16
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-speech==2.30.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.24.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
Js2Py==0.74
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
pipwin==0.5.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pyjsparser==2.7.1
pyparsing==3.2.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pyttsx3==2.98
pywin32==308
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.0.7
simpleaudio==1.0.4
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
